* moglcore: Make 64-bit clean, part I
  - Core functions and manually defined OpenGL functions
    in gl_manual.c updated to become 64-bit safe.

  - Associated M-file wrappers updated accordingly.

  => Wrap pointers in double's as opaque 64-bit containers,
     as we do in other places, whenever communicating with
     the scripting environment.

  => Use size_t and range checks in internal code to handle
     32 bit / 64 bit architectures.

  => moglcore for OS/X Intel Octave/Matlab rebuilt.

* moglcore: Prepare 64-bit support for auto-generated
  code in gl_auto.c

  - The automatic code generator scripts are prepared to
    generate 64-bit safe pointer wrapping code, but not yet
    tested.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@1992 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ Psychtoolbox/PsychOpenGL/MOGL/wrap/glGetBufferPointerv.m
@@ -7,6 +7,7 @@ function params = glGetBufferPointerv( target, pname )
 % C function:  void glGetBufferPointerv(GLenum target, GLenum pname, GLvoid** params)
 
 % 02-Jan-2006 -- created (RFM)
+% 24-Mar-2011 -- Made 64-bit clean (MK)
 
 % ---allocate---
 % ---protected---
@@ -15,7 +16,7 @@ if nargin~=2,
     error('invalid number of arguments');
 end
 
-params=uint32(0);
+params=uint64(0);
 moglcore( 'glGetBufferPointerv', target, pname, params );
 
 return
