+++ Psychtoolbox/PsychOpenGL/MOGL/wrap/glGetPointerv.m
@@ -7,6 +7,7 @@ function params = glGetPointerv( pname )
 % C function:  void glGetPointerv (GLenum pname, GLvoid* *params);
 
 % 01-Jan-2006 -- created (RFM)
+% 24-Mar-2011 -- Made 64-bit clean (MK)
 
 % ---allocate---
 % ---protected---
@@ -15,7 +16,7 @@ if nargin~=1,
     error('invalid number of arguments');
 end
 
-params=uint32(0);
+params=uint64(0);
 moglcore( 'glGetPointerv', pname, params );
 
 return
