+++ Psychtoolbox/PsychOpenGL/MOGL/wrap/glGetVertexAttribPointerv.m
@@ -7,6 +7,7 @@ function params = glGetVertexAttribPointerv( index, pname )
 % C function:  void glGetVertexAttribPointerv (GLuint index, GLenum pname, GLvoid** pointer)
 
 % 02-Jan-2006 -- created (RFM)
+% 24-Mar-2011 -- Made 64-bit clean (MK)
 
 % ---allocate---
 % ---protected---
@@ -15,7 +16,7 @@ if nargin~=2,
     error('invalid number of arguments');
 end
 
-params=uint32(0);
+params=uint64(0);
 moglcore( 'glGetVertexAttribPointerv', index, pname, params );
 
 return
